updated dispatch and selector to use hooks, moved SignupModal into components folder
